requirements.txt
# Data processing
requests==2.31.0
tqdm==4.66.1
python-dotenv==1.0.0

# OpenSearch
opensearch-py==2.4.2

# Testing
pytest==7.4.3

# Utilities
python-dateutil==2.8.2

# FastAPI 
fastapi==0.104.1

# Uvicorn 
uvicorn==0.24.0

# Pydantic 
pydantic==2.5.2

# MCP SDK -  Anthropic's Model Context Protocol
mcp[cli]>=1.0.0

# Sentence Transformers - for generating text embeddings (vector search)
sentence-transformers==2.2.2

# HTTPX - async HTTP client (needed for MCP and FastAPI)
httpx==0.25.2

# Anthropic SDK - to call Claude
anthropic==0.39.0